requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudevents==1.10.1
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
deprecation==2.1.0
distro==1.9.0
dnspython==2.4.2
exceptiongroup==1.2.0
Flask==3.0.0
fonttools==4.42.1
frozenlist==1.4.0
functions-framework==3.5.0
googleapis-common-protos==1.62.0
grpc-gateway-protoc-gen-openapiv2==0.1.0
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
loguru==0.7.2
lz4==4.3.2
MarkupSafe==2.1.3
matplotlib==3.7.2
multidict==6.0.4
nextcord==2.5.0
numpy==1.25.2
openai==1.6.1
packaging==23.1
pandas==2.1.0
Pillow==10.0.0
pinecone-client==2.2.4
plotly==5.16.1
protobuf==3.20.3
pydantic==2.5.3
pydantic_core==2.14.6
pymongo==4.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
sniffio==1.3.0
tenacity==8.2.3
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
watchdog==3.0.0
Werkzeug==3.0.1
win32-setctime==1.1.0
yarl==1.9.2
